REFACTOR(#5): make utils.logging a seprated py pkg
